// src/server/api/models/insurance/models/Claim.ts
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

import type { ClaimStatus } from './ClaimStatus';
import type { InsuranceProductType } from './InsuranceProductType';
import type { LocalDate } from './LocalDate';
import type { LocalDateTime } from './LocalDateTime';
import type { UUID } from './UUID';

export type Claim = {
    /**
     * The unique ID for the record. This is the ID that is used in all services and related records when referencing the record.
     */
    id: UUID;
    /**
     * The ID of the user that created the record.
     */
    createdBy: string;
    /**
     * When the record was created.
     */
    createdAt: LocalDateTime;
    /**
     * The ID of the user that last modified the record.
     */
    updatedBy: string;
    /**
     * When the record was last modified.
     */
    updatedAt: LocalDateTime;
    /**
     * Indicates if this is a deleted record
     */
    deleted: boolean;
    /**
     * The client ID from the crm service
     */
    clientId: UUID;
    /**
     * The client description
     */
    clientLabel: string;
    /**
     * The client code from CRM
     */
    clientCode: string;
    /**
     * A unique claim number
     */
    claimNo: string;
    /**
     * A claim number generated by the insurance partner
     */
    refClaimNo: string;
    /**
     * The claim description
     */
    claimLabel: string;
    /**
     * The claim currency
     */
    currency: string;
    /**
     * Notes related to the record
     */
    notes?: string | null;
    /**
     * The status of the claim
     */
    status: ClaimStatus;
    /**
     * The amount claimed
     */
    claimAmount: number | null;
    /**
     * The approved amount for the claim
     */
    approvedAmount?: number | null;
    /**
     * Date that the claim was submitted
     */
    claimDate: LocalDate;
    /**
     * Date that the claim was processed
     */
    dateProcessed?: LocalDate | null;
    /**
     * The policy ID
     */
    policyId: UUID | null;
    /**
     * The policy label
     */
    policyLabel: string | null;
    /**
     * The policy number
     */
    policyNo: string | null;
    /**
     * The type of the policy
     */
    productType: InsuranceProductType;
};